Improved - Modulo Agenda
Si la persona no tiene deuda y anula la visita, si le da la opción de borrarlo de la agenda. De lo contrario, le mostrará un mensaje que impide este proceso de eliminación.

+++ app/Http/Controllers/admin/schedules/VisitController.php
@@ -166,8 +166,17 @@ class VisitController extends Controller
     {
         try {
             $this->authorize('delete', $visita);
+            #validar si cliente tiene deudas 
+            if($visita->customer->haveDebtsCustomer())
+            {
+                #No se puede eliminar
+                return response()->json([
+                    'success' => false,
+                    'message' => "No se pudo eliminar la visita debido a una deuda pendiente del cliente"
+                ]);   
+            }
+            #client sin deudas
             $visita->delete();
-            
             return response()->json([
                 'success' => true,
                 'message' => "La visita ha sido eliminada con éxito"
